TPIII/Hit3/backend/requirements.txt
fastapi
pika
redis>=4.5.0
python-multipart
pillow
numpy
scipy
asyncio
uvicorn[standard]
